PLANNER-1058 mvn -DwithoutGwt profile + rename o-e-r-client to o-e-r-restclient
