+++ optashift-employee-rostering-webapp/pom.xml
@@ -36,24 +36,19 @@
       <groupId>org.optaplanner.openshift</groupId>
       <artifactId>optashift-employee-rostering-server</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.optaplanner.openshift</groupId>
-      <artifactId>optashift-employee-rostering-gwtui</artifactId>
-      <type>war</type>
-      <scope>runtime</scope>
-    </dependency>
+    <!-- The dependency optashift-employee-rostering-gwtui is added by the maven profile withGwt -->
+
+    <!-- JEE -->
     <dependency>
       <groupId>javax</groupId>
       <artifactId>javaee-api</artifactId>
       <scope>provided</scope>
     </dependency>
-
     <!-- Swagger -->
     <dependency>
       <groupId>io.swagger</groupId>
       <artifactId>swagger-jaxrs</artifactId>
     </dependency>
-
     <!-- Logging -->
     <dependency>
       <groupId>org.slf4j</groupId>
@@ -63,7 +58,7 @@
     <!-- Testing -->
     <dependency>
       <groupId>org.optaplanner.openshift</groupId>
-      <artifactId>optashift-employee-rostering-client</artifactId>
+      <artifactId>optashift-employee-rostering-restclient</artifactId>
       <scope>test</scope>
     </dependency>
     <dependency>
@@ -153,6 +148,23 @@
 
   <profiles>
     <profile>
+      <id>withGwt</id>
+      <activation>
+        <property>
+          <name>!withoutGwt</name>
+        </property>
+      </activation>
+      <dependencies>
+        <dependency>
+          <groupId>org.optaplanner.openshift</groupId>
+          <artifactId>optashift-employee-rostering-gwtui</artifactId>
+          <type>war</type>
+          <scope>runtime</scope>
+        </dependency>
+      </dependencies>
+    </profile>
+
+    <profile>
       <!-- Use this profile for any OpenShift specific customization your app will need. -->
       <!-- By default that is to put the resulting archive into the 'deployments' folder. -->
       <id>openshift</id>
@@ -170,10 +182,27 @@
         </plugins>
       </build>
     </profile>
+
+    <!-- Skip entire Cargo execution and tests -->
     <profile>
-      <id>wildfly11</id>
+      <id>skipCargo</id>
       <activation>
-        <activeByDefault>true</activeByDefault>
+        <property>
+          <name>skipTests</name>
+          <value>true</value>
+        </property>
+      </activation>
+      <properties>
+        <cargo.maven.skip>true</cargo.maven.skip>
+      </properties>
+    </profile>
+    
+    <profile>
+      <id>cargoWildfly11</id>
+      <activation>
+        <property>
+          <name>!cargo.maven.skip</name>
+        </property>
       </activation>
       <properties>
         <cargo.container.id>wildfly11x</cargo.container.id>
@@ -209,24 +238,11 @@
 
     <!-- Start the container and don't run tests -->
     <profile>
-      <id>run</id>
+      <id>runCargo</id>
       <properties>
         <cargo.goal>run</cargo.goal>
       </properties>
     </profile>
 
-    <!-- Skip entire Cargo execution and tests -->
-    <profile>
-      <id>skip-cargo</id>
-      <activation>
-        <property>
-          <name>skipTests</name>
-          <value>true</value>
-        </property>
-      </activation>
-      <properties>
-        <cargo.maven.skip>true</cargo.maven.skip>
-      </properties>
-    </profile>
   </profiles>
 </project>
