+++ optashift-employee-rostering-webapp/src/main/webapp/index.html
@@ -6,10 +6,9 @@
 <body>
   <h1>What web UI technology would you like to see?</h1>
   <ul>
-    <li><a href="gwtui/gwtui.html">GWT user interface</a></li>
+    <li><a href="gwtui/gwtui.html">GWT user interface</a> (not available if build with <code>mvn -DwithoutGwt</code>)</li>
     <li>Angular user interface (not yet implemented)</li>
     <li><a href="swagger/index.html">REST API documentation</a></li>
   </ul>
-  <p>You can easily delete the web UI's you don't use in your fork without impacting the server.</p>
 </body>
 </html>
