+++ optashift-employee-rostering-webapp/src/main/webapp/swagger/swagger-ui.js
